Move objects API to separate package
