+++ web/src/collection/state/reducers.js
@@ -7,7 +7,7 @@ import taskReducer from "./tasks/reducer";
 import taskCacheReducer from "./taskCache/reducer";
 import taskLogsReducer from "./taskLogs/reducer";
 import templateReducer from "./templates/reducer";
-import objectCacheReducer from "./objectCache/reducer";
+import objectCacheReducer from "../../application/objects/state/objectCache/reducer";
 import presetReducer from "./presets/reducer";
 import fileExclusionsCacheReducer from "../../application/file-exclusion/state/reducer";
