+++ web/src/collection/state/selectors.js
@@ -53,6 +53,11 @@ export const selectTaskLogs = (state) => selectColl(state).taskLogs;
 export const selectTemplates = (state) => selectColl(state).templates;
 
 /**
+ * Select object cache.
+ */
+export const selectObjectCache = (state) => selectColl(state).objectCache;
+
+/**
  * Select cached file objects.
  */
 export const selectCachedObjects = (fileId) => (state) =>
